고장 수정

+++ src/main/java/com/korit/dreampath_back/entity/User.java
@@ -31,7 +31,7 @@ public class User {
     private String oAuth2Provider;
 
     private int remaining;
-    private int starPoint;
+    private double starPoint;
     private int remainPoint;
 
     private int accountExpired;
